+++ gulpfile.js
@@ -3,9 +3,7 @@
 /**
  * [Gulp](http://gulpjs.com/) plugin for [happiness](https://github.com/JedWatson/happiness)
  * @module gulp-happiness
- *
  * @author Oleg Dutchenko <dutchenko.o.wezom@gmail.com>
- * @version 0.0.4
  */
 
 // ----------------------------------------
@@ -21,16 +19,11 @@ const gulpHappiness = require('./index');
 
 gulp.task('lint', function () {
 	let sources = [
-		'./*.js',
-		'./node_modules/gulp-not-supported-file/*.js'
+		'./*.js'
 	];
 
 	return gulp.src(sources)
-		.pipe(gulpHappiness({
-			noUnderscore: false
-		}))
-		.pipe(gulpHappiness.format({
-			showHappyFiles: true,
-			noUnderscore: false
-		}));
+		.pipe(gulpHappiness())
+		.pipe(gulpHappiness.format())
+		.pipe(gulpHappiness.failAfterError());
 });