+++ index.js
@@ -28,8 +28,17 @@ const notSupportedFile = require('gulp-not-supported-file');
  * Saved plug-in name for use in terminal logs
  * @const {string}
  * @private
+ * @sourceCode
  */
 const pluginName = 'gulp-happiness';
+
+/**
+ * Text message about how get more information
+ * @const {string}
+ * @private
+ * @sourceCode
+ */
+const moreInfo = '\n    Info:\n    Use gulpHappiness.format() method for more information about errors';
 
 /**
  * Plugin error constructor
@@ -43,7 +52,38 @@ function pluginError (sample, options) {
 	return new gutil.PluginError(pluginName, sample, options);
 }
 
-function getEslintData (file, pluginError, runOptions) {
+/**
+ * Get 'error' or 'errors' text
+ * @param {number} count
+ * @return {string}
+ * @private
+ * @sourceCode
+ */
+function getErrorText (count) {
+	return count > 1 ? 'errors' : 'error';
+}
+
+/**
+ * Get 'path' or 'paths' text
+ * @param {number} count
+ * @return {string}
+ * @private
+ * @sourceCode
+ */
+function getPathText (count) {
+	return count > 1 ? 'paths' : 'path';
+}
+
+/**
+ * Get eslint result data from file
+ * @param {File} file
+ * @param {function} pluginError
+ * @param {Object} [runOptions={}]
+ * @return {Object|Array}
+ * @private
+ * @sourceCode
+ */
+function getEslintData (file, pluginError, runOptions = {}) {
 	let notSupported = notSupportedFile(file, pluginError, {
 		silent: runOptions.silent,
 		noUnderscore: runOptions.noUnderscore,
@@ -80,12 +120,12 @@ function getEslintData (file, pluginError, runOptions) {
  * @param {boolean} [options.noUnderscore=true]
  * @param {boolean} [options.noEmpty=true]
  * @returns {DestroyableTransform} through2.obj
+ * @sourceCode
  */
 function gulpHappiness (options = {}) {
 	let runOptions = _cloneDeep(options);
 
-	return through2.obj(function (file, ...args) {
-		let cb = args[1];
+	return through2.obj(function (file, enc, cb) {
 		let notSupported = notSupportedFile(file, pluginError, {
 			silent: runOptions.silent,
 			noUnderscore: runOptions.noUnderscore,
@@ -97,7 +137,14 @@ function gulpHappiness (options = {}) {
 			return cb(...notSupported);
 		}
 
-		happiness.lintText(String(file.contents), runOptions, function (err, data) {
+		let lintOptions = _cloneDeep(runOptions.linter || {});
+		let fixProblems = runOptions.fix;
+
+		if (fixProblems) {
+			happiness.eslintConfig.fix = true;
+		}
+
+		happiness.lintText(String(file.contents), lintOptions, function (err, data) {
 			if (err) {
 				return cb(pluginError(err));
 			}
@@ -105,6 +152,15 @@ function gulpHappiness (options = {}) {
 			data.results.forEach(result => {
 				result.filePath = file.path;
 			});
+
+			if (fixProblems) {
+				let output = data.results[0].output;
+				if (data.errorCount === 0 && typeof output === 'string') {
+					file.contents = Buffer.from(output, enc);
+				} else {
+					console.log(gutil.colors.yellow(`\nCannot auto fix ${file.path}\nDo it yourself manual\n`));
+				}
+			}
 
 			file.eslint = data;
 			cb(null, file);
@@ -117,8 +173,11 @@ function gulpHappiness (options = {}) {
  * @param {string|Object} [formatter='default'] if it is Object using as options
  * @param {Object}        [options={}]
  * @param {boolean}       [options.silent]
+ * @param {boolean}       [options.noUnderscore=true]
+ * @param {boolean}       [options.noEmpty=true]
  * @param {boolean}       [options.showHappyFiles]
  * @returns {DestroyableTransform} through2.obj
+ * @sourceCode
  */
 gulpHappiness.format = function (formatter = 'default', options = {}) {
 	if (_isPlainObject(formatter)) {
@@ -165,6 +224,7 @@ gulpHappiness.format = function (formatter = 'default', options = {}) {
 
 		if (_isFunction(formatter)) {
 			formatter(eslintData, runOptions);
+			file.eslintIsFormeated = true;
 			return cb(null, file);
 		}
 
@@ -187,13 +247,113 @@ gulpHappiness.format = function (formatter = 'default', options = {}) {
 
 				console.log(gutil.colors.red(`SAD FILE > ${file.path}`));
 				console.log(result);
+				file.eslintIsFormeated = true;
 				return cb(null, file);
 			} catch (err) {
 				return cb(pluginError(err));
 			}
 		}
 
-		return cb(pluginError(`No suitable formatter - ${formatter}`));
+		return cb(pluginError(`Error! No suitable formatter - ${formatter}`));
+	});
+};
+
+/**
+ * Failing if file has eslint errors
+ * @param {Object}  [options={}]
+ * @param {boolean} [options.silent]
+ * @param {boolean} [options.noUnderscore=true]
+ * @param {boolean} [options.noEmpty=true]
+ * @returns {DestroyableTransform} through2.obj
+ * @sourceCode
+ */
+gulpHappiness.failOnError = function (options = {}) {
+	let runOptions = _cloneDeep(options);
+
+	return through2.obj(function (file, ...args) {
+		let cb = args[1];
+		let filePaths = [];
+		let eslintData = getEslintData(file, pluginError, runOptions);
+
+		if (Array.isArray(eslintData)) {
+			eslintData.shift();
+			return cb(...eslintData);
+		}
+
+		if (eslintData.errorCount === 0) {
+			return cb(null, file);
+		}
+
+		eslintData.results.forEach(result => {
+			filePaths.push(result.filePath);
+		});
+
+		let count = eslintData.errorCount;
+		let errorText = getErrorText(count);
+		let pathText = getPathText(filePaths.length);
+		let errorMsg = `Fail on Error! ${count} ${errorText} in ${pathText}:\n    ${filePaths.join('\n    ')}`;
+
+		if (file.eslintIsFormeated !== true) {
+			errorMsg += moreInfo;
+		}
+
+		return cb(pluginError(errorMsg));
+	});
+};
+
+/**
+ * Look after checking all the streamed files,
+ * and if at least one of them has errors it will fail.
+ * __Note!__ This method does not transfer files further to the stream!
+ * @param {Object}  [options={}]
+ * @param {boolean} [options.silent]
+ * @param {boolean} [options.noUnderscore=true]
+ * @param {boolean} [options.noEmpty=true]
+ * @returns {DestroyableTransform} through2.obj
+ * @sourceCode
+ */
+gulpHappiness.failAfterError = function (options = {}) {
+	let runOptions = _cloneDeep(options);
+	let eslintIsFormatted = false;
+	let allErrorsCount = 0;
+	let filePaths = [];
+
+	return through2.obj(function (file, ...args) {
+		let cb = args[1];
+		let eslintData = getEslintData(file, pluginError, runOptions);
+
+		if (Array.isArray(eslintData)) {
+			eslintData.shift();
+			return cb(...eslintData);
+		}
+
+		if (eslintData.errorCount === 0) {
+			return cb();
+		}
+
+		eslintIsFormatted = file.eslintIsFormeated;
+		allErrorsCount += eslintData.errorCount;
+		eslintData.results.forEach(result => {
+			let count = result.errorCount;
+			let errorText = getErrorText(count);
+
+			filePaths.push(`has ${count} ${errorText} in ${result.filePath}`);
+		});
+
+		cb();
+	}, function (cb) {
+		if (allErrorsCount === 0) {
+			return cb();
+		}
+
+		let errorText = getErrorText(allErrorsCount);
+		let pathText = getPathText(filePaths.length);
+		let errorMsg = `Fail after Error! ${allErrorsCount} ${errorText} in ${pathText}:\n    ${filePaths.join('\n    ')}`;
+
+		if (eslintIsFormatted !== false) {
+			errorMsg += moreInfo;
+		}
+		cb(pluginError(errorMsg));
 	});
 };