+++ CHANGELOG.md
@@ -1,14 +1,24 @@
 # Change log
+
+:back: [README.md](./README.md)
+
 > _All notable changes to this project will be documented in this file._  
 > _This project adheres to [Semantic Versioning](http://semver.org/)._
 
+
+### [0.2.0] - 2017-05-26
+
+#### Added
+
+Public methods
+- gulpHappiness.format()
+- gulpHappiness.failOnError()
+- gulpHappiness.failAfterError()
 
 ### [0.1.0] - 2017-05-22
 
 #### Changed
 - dependencies and devDependencies
-- preparing for npm tests
-- preparing for [Travis CI](https://travis-ci.org/dutchenkoOleg/gulp-not-supported-file) builds
 - `index.js` full rewrite
 
 ---